rename_fragments_dir
