+++ app/src/main/res/layout/fragment_test.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.TestFragment" >
+    tools:context=".fragment.TestFragment" >
 
     <TextView
         android:id="@+id/testTextView"
