+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,14 +7,14 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.nje.mobileszkozokprojekt.fragments.HomeFragment"
+        android:name="com.nje.mobileszkozokprojekt.fragment.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home"
         />
 
     <fragment
         android:id="@+id/testFragment"
-        android:name="com.nje.mobileszkozokprojekt.fragments.TestFragment"
+        android:name="com.nje.mobileszkozokprojekt.fragment.TestFragment"
         android:label="Test"
         tools:layout="@layout/fragment_test"
         />
